Move more GraphQL requests
